Some architecture cleanup
