+++ src/data/feedbackMap.js
@@ -1,0 +1,23 @@
+export default {
+  DEFAULT: {
+    message: "Sign up to get notified when it's ready",
+    className: "feedback",
+  },
+  SUCCESS: {
+    message: "Thanks for subscribing!",
+    className: "feedback feedback--success",
+  },
+  INVALID_EMAIL: {
+    message: "That email address is already subscribed",
+    className: "feedback feedback--error",
+  },
+  EMAIL_EXISTS: {
+    message: "That email address is already subscribed",
+    className: "feedback feedback--error",
+  },
+  ERROR: {
+    message:
+      "There was an error subscribing. Refresh the page or try again later",
+    className: "feedback feedback--error",
+  },
+}
